// lib/types/sanity.types.ts
export type CaseStudies = {
  _type: 'caseStudies';
  caseStudies: number;
  orderRank: 'latest' | 'random';
};

export type GoogleReviews = {
  _type: 'googleReviews';
  reviews: number;
};

export type InfoTable = {
  _type: 'infoTable';
  title?: string;
  description?: string;
  firstRowHeader?: boolean;
  infoTable: Table;
};

export type VideoEmbed = {
  _type: 'videoEmbed';
  video?: YoutubeVideo;
  autoplay?: boolean;
  controls?: boolean;
};

export type RawHtml = {
  _type: 'rawHtml';
  rawHtml: string;
  centered?: boolean;
};

export type VerticalTabsItem = {
  _type: 'verticalTabsItem';
  title: string;
  content: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h2' | 'h3' | 'h4' | 'blockquote' | 'hardBreak';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
};

export type VerticalTabs = {
  _type: 'verticalTabs';
  tabs: Array<
    {
      _key: string;
    } & VerticalTabsItem
  >;
};

export type TwoSlotRichTextImage = {
  _type: 'twoSlotRichTextImage';
  leftSlotType: 'richText' | 'image';
  leftRichText?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  leftImage?: {
    id?: string;
    url?: string;
    alt_text?: string;
  };
  leftImageCaption?: string;
  leftImageAttribution?: string;
  rightSlotType: 'richText' | 'image';
  rightRichText?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  rightImage?: {
    id?: string;
    url?: string;
    alt_text?: string;
  };
  rightImageCaption?: string;
  rightImageAttribution?: string;
};

export type TwoColumnRichTextAndForm = {
  _type: 'twoColumnRichTextAndForm';
  richText: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h2' | 'h3' | 'h4' | 'blockquote' | 'hardBreak';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  formHeading?: string;
  rentalSpaceReference: string;
  swapSides?: boolean;
};

export type TwoColumnRichTextAndImage = {
  _type: 'twoColumnRichTextAndImage';
  richText: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h2' | 'h3' | 'h4' | 'blockquote' | 'hardBreak';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  image?: {
    id?: string;
    url?: string;
    alt_text?: string;
  };
  swapSides?: boolean;
};

export type TwoColumnRichText = {
  _type: 'twoColumnRichText';
  richText: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h2' | 'h3' | 'h4' | 'blockquote' | 'hardBreak';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
};

export type TwoColumnList = {
  _type: 'twoColumnList';
  title?: string;
  items: Array<string>;
  showBackground?: boolean;
};

export type TwoCardSection = {
  _type: 'twoCardSection';
  cards: Array<
    {
      _key: string;
    } & ContentCard
  >;
};

export type Spacer = {
  _type: 'spacer';
  size: 'spacer-small' | 'spacer-medium' | 'spacer-large';
};
// ...
export type PropertiesCta = {
  _id: string;
  _type: 'propertiesCta';
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title: string;
  content: string;
  ctaLink: string;
};

export type PlainText = {
  _type: 'plainText';
  title: string;
  body?: string;
};

export type PdfDownload = {
  _type: 'pdfDownload';
  pdf?: {
    text?: string;
    href: string;
  };
};
// ...
export type LinkList = {
  _type: 'linkList';
  title?: string;
  items: Array<{
    title: string;
    url: string;
    _key: string;
  }>;
};

export type IndexHeroes = {
  _id: string;
  _type: 'indexHeroes';
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  indexHeroImages?: {
    featuredLocations?: BrandfolderImage;
    findYourSpace?: BrandfolderImage;
    newsAndReports?: BrandfolderImage;
    caseStudies?: BrandfolderImage;
  };
};

export type ImageRow = {
  _type: 'imageRow';
  image?: {
    id?: string;
    url?: string;
    alt_text?: string;
  };
  title: string;
  content?: string;
  ctaText: string;
  ctaUrl: string;
  ctaColour?: 'navy' | 'teal' | 'orange' | 'white';
  ctaIcon?: 'phone' | 'email' | 'search';
  cta2text?: string;
  cta2url?: string;
  cta2colour?: 'navy' | 'teal' | 'orange' | 'white';
  cta2icon?: 'phone' | 'email' | 'search';
  dark?: boolean;
  overlayText: boolean;
};

export type HorizontalContentCardSection = {
  _type: 'horizontalContentCardSection';
  cards?: Array<
    {
      _key: string;
    } & HorizontalContentCard
  >;
};

export type HorizontalContentCard = {
  _type: 'horizontalContentCard';
  image?: {
    id?: string;
    url?: string;
  };
  title: string;
  contentType?: 'text' | 'block';
  content?: string;
  richText?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h2' | 'h3' | 'h4' | 'blockquote' | 'hardBreak';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  ctaType?: 'link' | 'button' | 'clickable' | 'none';
  linkText?: string;
  linkUrl?: string;
  swapSides?: boolean;
};
// ...
export type ThreeCardSection = {
  _type: 'threeCardSection';
  cards: Array<
    {
      _key: string;
    } & ContentCard
  >;
};

export type FullWidthImage = {
  _type: 'fullWidthImage';
  image?: {
    id?: string;
    url?: string;
    alt_text?: string;
  };
  fullHeight?: boolean;
};

export type FourCardSection = {
  _type: 'fourCardSection';
  cards: Array<
    {
      _key: string;
    } & ContentCard
  >;
};

export type FooterLink = {
  _type: 'footerLink';
  label: string;
  href: string;
};

export type Footer = {
  _type: 'footer';
  featuredLinks?: Array<
    {
      _key: string;
    } & FooterLink
  >;
};

export type FindYourSpace = {
  _id: string;
  _type: 'findYourSpace';
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  metadata?: Metadata;
  title: string;
  slug: Slug;
  heroImage?: BrandfolderImage;
  summary: string;
  linkText?: string;
  pageIntroText: RichText;
  pageIntroCTATitle: string;
  pageIntroCTADescription: string;
  pageIntroCTALinkText: string;
  pageIntroCTALinkUrl: string;
  content: Array<
    | ({
        _key: string;
      } & PlainText)
    | ({
        _key: string;
      } & RichText)
    | ({
        _key: string;
      } & CtaRichText)
    | ({
        _key: string;
      } & TwoColumnRichText)
    | ({
        _key: string;
      } & TwoCardSection)
    | ({
        _key: string;
      } & ThreeCardSection)
    | ({
        _key: string;
      } & FourCardSection)
    | ({
        _key: string;
      } & ImageRow)
    | ({
        _key: string;
      } & HorizontalContentCard)
    | ({
        _key: string;
      } & TwoColumnList)
    | ({
        _key: string;
      } & TwoColumnRichTextAndImage)
    | ({
        _key: string;
      } & TwoColumnRichTextAndForm)
    | ({
        _key: string;
      } & TwoSlotRichTextImage)
    | ({
        _key: string;
      } & FullWidthImage)
    | ({
        _key: string;
      } & VerticalTabs)
    | ({
        _key: string;
      } & PdfDownload)
    | ({
        _key: string;
      } & Spacer)
    | ({
        _key: string;
      } & RawHtml)
    | ({
        _key: string;
      } & VideoEmbed)
    | ({
        _key: string;
      } & InfoTable)
    | ({
        _key: string;
      } & GoogleReviews)
    | ({
        _key: string;
      } & CaseStudies)
  >;
  sideContent?: Array<
    | ({
        _key: string;
      } & ContentCard)
    | ({
        _key: string;
      } & RichText)
    | ({
        _key: string;
      } & LinkList)
  >;
  featuredProperties?: FeaturedProperties;
  orderRank?: string;
};
// ...
export type FeaturedProperties = {
  _type: 'featuredProperties';
  title?: string;
  properties?: Array<string>;
};
// ...
export type ContentCard = {
  _type: 'contentCard';
  image?: {
    id?: string;
    url?: string;
    alt_text?: string;
  };
  isFeatured?: boolean;
  title: string;
  contentType?: 'text' | 'block';
  content?: string;
  richText?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h2' | 'h3' | 'h4' | 'blockquote' | 'hardBreak';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  ctaType?: 'link' | 'button' | 'clickable' | 'none';
  linkText?: string;
  linkUrl?: string;
  cta2Type?: 'link' | 'button';
  link2Text?: string;
  link2Url?: string;
};

export type CtaRichText = {
  _type: 'ctaRichText';
  pageIntroText: RichText;
  pageIntroCTATitle: string;
  pageIntroCTADescription: string;
  pageIntroCTALinkText: string;
  pageIntroCTALinkUrl: string;
};

export type RichText = {
  _type: 'richText';
  richText: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: 'span';
      _key: string;
    }>;
    style?: 'normal' | 'h2' | 'h3' | 'h4' | 'blockquote' | 'hardBreak';
    listItem?: 'bullet' | 'number';
    markDefs?: Array<{
      href?: string;
      _type: 'link';
      _key: string;
    }>;
    level?: number;
    _type: 'block';
    _key: string;
  }>;
  centered?: boolean;
};
// ...
export type Slug = {
  _type: 'slug';
  current: string;
  source?: string;
};

export type BrandfolderPdf = {
  _type: 'brandfolderPdf';
  id?: string;
  url?: string;
};

export type AboutUs = {
  _id: string;
  _type: 'aboutUs';
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  metadata?: Metadata;
  title: string;
  pageIntroText: string;
  heroImage?: BrandfolderImage;
  content?: ComposableContent;
};

export type BrandfolderImage = {
  _type: 'brandfolderImage';
  id?: string;
  url?: string;
  alt_text?: string;
  width?: number;
  height?: number;
};

export type Metadata = {
  _type: 'metadata';
  title?: string;
  description?: string;
  noIndex?: boolean;
  noFollow?: boolean;
};

export type YoutubeVideo = {
  _type: 'youtubeVideo';
  id: string;
  title?: string;
  description?: string;
  publishedAt?: string;
  thumbnails?: Array<string>;
};

export type Table = {
  _type: 'table';
  rows?: Array<
    {
      _key: string;
    } & TableRow
  >;
};